+++ app/sandwichingredients/serializers/category.py
@@ -1,0 +1,11 @@
+from rest_framework import serializers
+
+from ..models import Category
+
+
+class CategorySerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Category
+        fields = (
+            'name',
+        )
